backend/requirements.txt
# FastAPI и основные зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.36  # Updated for Python 3.13 compatibility
alembic==1.14.0  # Updated for compatibility with SQLAlchemy 2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7
python-dotenv==1.0.0

# API Clients
httpx==0.25.2
openai==1.3.7

# Background Jobs
rq==1.15.1
redis==5.2.1  # Updated for Redis 7.2+ server compatibility

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
xlsxwriter==3.1.9
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Monitoring
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-httpx==0.24.0